backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
psycopg2-binary==2.9.7
marshmallow==3.20.1
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0 